python-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai==1.51.0
pydantic==2.8.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
httpx==0.27.2
httpcore==1.0.6

google-genai==1.46.0
